Implement Market Making/Taking strategy for Rainforest Resin
This code is based off of [linear utility's code](https://github.com/ericcccsliu/imc-prosperity-2)

We first take orders from bots that are favourable. i.e. their bids >
our `fair_price` or their asks < our `fair_price`

We then set bid at `fair_value - spread` ask at `fair_value + spread` to
make a market.

Due to `position_limit` we want to set the biggest volume to be within
the limits.

For example, if the limit is 50 and our net `position` from the previous
tick was -2. Suppose that we took 20 bids. Then our current position `cpos`
is now -22. We can therefore place a sell order with volume at most
28.

Whether your bids or asks are over the limit are determined
independently. So you cannot place 1000 bids and 1000 asks to cancel
out to stay within the limits.

Continuing with the example, suppose we took 0 asks. Then our current
position `cpos` is still `2` and so we can place a bid order of volume
48.

When `position == position_limit` we are not allowed to buy any more
units until our position is reduced. This applies to sell mutatis
mutandis. jWe want to avoid having a position equal to plus or minus
`position_limit` because of the opportunity cost with not being able to
trade in a certain direction.

We make the additional improvement where if `position ==
-position_limit` or `position == position_limit`, then we are willing to
put up a tighter spread in the direction we want to liquidate to
incentivise bots to clear a bit of our excess position.